ai-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
faster-whisper==1.1.0
insightface==0.7.3
onnxruntime==1.20.1
mediapipe==0.10.18
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
langchain-google-genai==2.0.7
langchain-core==0.3.28
google-generativeai==0.8.3
pydantic==2.10.3
redis[hiredis]==5.2.1
python-dotenv==1.0.1
httpx==0.28.1
